requirements.txt
python-dotenv
requests>=2.24.0
loguru>=0.6.0
python-gitlab
anthropic
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.0
python-gitlab>=3.0.0
loguru>=0.7.0